+++ src/com/intellij/idea/plugin/hybris/impex/highlighting/findUsages/HighlightMacrosHandlerFactory.kt
@@ -1,10 +1,11 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -48,6 +49,7 @@ class HighlightMacrosHandlerFactory : HighlightUsagesHandlerFactoryBase() {
 
         override fun computeUsages(targets: List<PsiElement>) {
             val name = getName()
+            // TODO: fix usages
             PsiTreeUtil.collectElements(myTarget.containingFile) {
                 (it is ImpexMacroNameDec && it.textMatches(name))
                     || (it is ImpexMacroUsageDec && it.reference?.resolve()?.text == name)