+++ src/com/intellij/idea/plugin/hybris/impex/search/ImpexReferenceSearcher.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2024 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -40,7 +41,7 @@ class ImpexReferenceSearcher : QueryExecutorBase<PsiReference, ReferencesSearch.
             && !ImpexPsiUtils.isMacroUsage(elementToSearch)
             && !ImpexPsiUtils.isMacroNameDeclaration(elementToSearch)
         ) return
-
+        //TODO: fix reference search
         PsiTreeUtil.collectElements(elementToSearch.containingFile)
         {
             it is ImpexMacroUsageDec